[feat]: Implement router RecruitEditView
