+++ lib/src/views/recruitment/recruitment_view.dart
@@ -34,25 +34,7 @@ class _RecruitViewState extends State<RecruitView> {
                 color: AppColorsTheme().gfGray400Color,
               ),
               onPressed: () {
-                Navigator.push(
-                  context,
-                  PageRouteBuilder(
-                    pageBuilder: (context, animation, secondaryAnimation) => RecruitEditView(),
-                    transitionsBuilder: (context, animation, secondaryAnimation, child) {
-                      const begin = Offset(0.0, 1.0);
-                      const end = Offset.zero;
-                      const curve = Curves.easeInOut;
-
-                      var tween = Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
-                      var offsetAnimation = animation.drive(tween);
-
-                      return SlideTransition(
-                        position: offsetAnimation,
-                        child: child,
-                      );
-                    },
-                  ),
-                );
+                context.go('/recruitedit');
               },
           ),
         ],
